smart_travel_agency/apps/tools-main/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
asyncpg